+++ src/src_V2/modules/auth/components/authInfoWrapper/AuthInfoWrapper.jsx
@@ -1,5 +1,5 @@
-import styles from './styles.module.css';
-import logo from '../../assets/Travely.png';
+import styles from './AuthInfoWrapper.module.css';
+import logo from '../../../../assets/Travely.png';
 
 export default function AuthInfoWrapper({ children }) {
   return (
