+++ src/src_V2/modules/auth/components/authWrapper/authPageWrapper.module.css
@@ -1,4 +1,4 @@
-@import '../../styles/colors.css';
+@import '../../../../styles/colors.css';
 
 .authWrapper {
   display: flex;
