+++ src/src_V2/modules/auth/components/congratulations/Congratulations.module.css
