+++ src/src_V2/modules/auth/components/verifyEmail/VerifyEmail.module.css
@@ -1,4 +1,4 @@
-@import '../../styles/colors.css';
+@import '../../../../styles/colors.css';
 
 .content {
   max-width: 396px;
